Added initial snapcraft for buiding snaps of GHDL
This is an initial attempt. I've tested that the `ghdl`
commandline utility works and will build out of the home directory.

I've added a slot for libghdl but I haven't tested a consumer
that uses that interface. I've confirmed that it does export the
connection:

```
$> snap connections ghdl
Interface  Plug       Slot                Notes
content    -          ghdl:libghdl3-1804  -
home       ghdl:home  :home               -
```